pyproject.toml
[project]
name = "nim-proxy"
version = "0.1.0"
description = "Minimal NVIDIA NIM proxy for Claude Code"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
